fix: replaying endpoints with query-params
This fixes replaying requests for endpoints with query params, although there is still an issue that only the first matching request will be replayed.

Assuming we have a POST endpoint that we sent a payload to, and assuming we have made multiple requests to this endpoint with different payloads then only the first request will be replayed.

It might make sense to extend the matching logic to also check the request body if it is a POST request.